feat: Refactor timetable form and update project structure
