+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\timeScheduler",
+    "\\timeScheduler\\bin",
+    "\\timeScheduler\\bin\\Debug"
+  ],
+  "SelectedNode": "\\timeScheduler\\obj",
+  "PreviewInSolutionExplorer": false
+}
